refactor field context
