+++ src/components/vm/fields/ReactField.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import React, { createContext, useState } from "react"
+import React from "react"
 import ReactDOM from "react-dom"
 import Blockly from "blockly"
 import JacdacProvider from "../../../jacdac/Provider"
@@ -10,44 +10,12 @@ import AppTheme from "../../ui/AppTheme"
 import { Box } from "@material-ui/core"
 import { BlockDefinition } from "../toolbox"
 import { assert } from "../../../../jacdac-ts/src/jdom/utils"
+import { ReactFieldProvider } from "./ReactFieldContext"
 
 declare module "blockly" {
     interface Block {
         getColourTertiary(): string
     }
-}
-
-export interface ReactFieldContextProps<T = any> {
-    value: T
-    onValueChange: (value: T) => void
-}
-
-export const ReactFieldContext = createContext<ReactFieldContextProps>({
-    value: undefined,
-    onValueChange: undefined,
-})
-ReactFieldContext.displayName = "ReactField"
-
-export function ReactFieldProvider(props: {
-    value: any
-    onValueChange?: (newValue: any) => any
-    children: ReactNode
-}) {
-    const {
-        children,
-        value: initialValue,
-        onValueChange: onFieldValueChange,
-    } = props
-    const [value, setValue] = useState<any>(initialValue)
-    const onValueChange = (newValue: any) => {
-        setValue(newValue)
-        onFieldValueChange?.(newValue)
-    }
-    return (
-        <ReactFieldContext.Provider value={{ value, onValueChange }}>
-            {children}
-        </ReactFieldContext.Provider>
-    )
 }
 
 export type ReactFieldJSON = any
